+++ sofa/arguments.lua
@@ -6,7 +6,10 @@ function arguments.parse()
   local parser = argparse()
     :name("sofa")
     :description("A command execution engine powered by rofi.")
-    :epilog("For more information see: https://github.com/f4z3r/sofa")
+    :epilog([[The configuration is set at ~/.config/sofa/config.yaml, and its location can be
+overwritten using the SOFA_CONFIG environment variable.
+
+For more information see: https://github.com/f4z3r/sofa]])
     :add_complete()
 
   parser:flag("-i --interactive", "Whether to get interactive commands.")